Remove duplicate AuthenticationManager definition
Previously SimpleUsernamePasswordAuthFilter had a separate reference to an
AuthenticationManager from the one defined in
UsernamePasswordAuthenticationFilter. This means there was an unused
variable in UsernamePasswordAuthenticationFilter. It also meant that
SimpleUsernamePasswordAuthFilter.setAuthenticationManager was setting an
AuthenticationManager that wasn't used for authenticating.

This removes the unnecessary AuthenticationMnaager and instead sets the
UsernamePasswordAuthenticationFilter instance.